tast-tests: add runtime var crash.testInProgFile and crash.prefix
crash.testInProgFile controls whether or not to create test-in-prog file.
crash.prefix allow a customized prefix to be writen to test-in-prog.

BUG=b:193461265

TEST=tast run -var=crash.prefix=cuj. <ip> example.RuntimeVars
TEST=tast run -var=crash.testInProgFile=false  <ip> example.RuntimeVars
TEST=fast_build.sh -T

+++ src/chromiumos/tast/local/crash/crash.go
@@ -14,6 +14,7 @@ import (
 	"path/filepath"
 	"regexp"
 	"sort"
+	"strconv"
 	"strings"
 	"time"
 
@@ -109,6 +110,20 @@ const (
 	// FilterInIgnoreAllCrashes is a value to put in the filter-in file if
 	// you wish to ignore all crashes that happen during a test.
 	FilterInIgnoreAllCrashes = "none"
+)
+
+var (
+	testInProgFileStrVar = testing.RegisterVarString(
+		"crash.testInProgFile", // The variable controls if file test-in-prog be created for the running test.
+		// Default value is true, create file test-in-prog by default.
+		// When set the var to "false", test-in-prog file will not be created for the running test.
+		"true",
+		"The variable that controls if test-in-prog file need to be created for the running test")
+
+	crashPrefixStrVar = testing.RegisterVarString(
+		"crash.prefix", // The variable prefixed to the test case name in test-in-prog file.
+		"",             // By default no prefix will be added.
+		"the string that will be prefixed to the test case name in test-in-prog file.")
 )
 
 // DefaultDirs returns all standard directories to which crashes are written.
@@ -485,10 +500,26 @@ func processRunning(procName string) (bool, error) {
 	return false, nil
 }
 
+// parseTestInProgFileStr Parses testInProgStrVar to a boolean value.
+// If the value of testInProgStrVar is unsupported by strconv.ParseBool(), it will return true.
+func parseTestInProgFileStr() bool {
+	testInProgFile, err := strconv.ParseBool(testInProgFileStrVar.Value())
+
+	//If any parse error happens, set the value to true.
+	if err != nil {
+		testInProgFile = true
+	}
+	return testInProgFile
+}
+
 // MarkTestInProgress writes |name| to |testInProgressPath|, indicating to crash_reporter
 // that the given test is in progress.
 func MarkTestInProgress(name string) error {
-	if err := ioutil.WriteFile(testInProgressPath, []byte(name), 0644); err != nil {
+	genTestInProgFile := parseTestInProgFileStr()
+	if !genTestInProgFile {
+		return nil
+	}
+	if err := ioutil.WriteFile(testInProgressPath, []byte(crashPrefixStrVar.Value()+name), 0644); err != nil {
 		return errors.Wrap(err, "failed to write in-progress test name")
 	}
 	return nil
@@ -496,6 +527,10 @@ func MarkTestInProgress(name string) error {
 
 // MarkTestDone removes the file indicating which test is running.
 func MarkTestDone() error {
+	genTestInProgFile := parseTestInProgFileStr()
+	if !genTestInProgFile {
+		return nil
+	}
 	if err := os.Remove(testInProgressPath); err != nil && !os.IsNotExist(err) {
 		return errors.Wrap(err, "failed to remove in-progress test name")
 	}
